refactor(#3274): extract object-runtime prototype-chain helpers (WAVE-B slice 3)
Third slice of the `ensureObjectRuntime` decomposition, stacked on slice 2.
Lifts the prototype-chain helper-build block (~320 LOC) VERBATIM into a new
sibling module `object-runtime-prototype.ts` (`buildObjectPrototypeHelpers`),
threading the (minimal) captured scope through `ObjectPrototypeHelperState`.

Helpers relocated: __getPrototypeOf, __object_create, __object_setPrototypeOf,
__isPrototypeOf.

Cleanest slice — only 6 captures + 2 constants via state, no cross-module
imports, no downstream-const entanglement.

object-runtime.ts: 6,607 -> 6,299 LOC; ensureObjectRuntime now ~3,493
(from ~7,378 pre-decomposition — ~53% extracted across slices 1-3).

Gates: `prove-emit-identity check` -> IDENTICAL (39/39 gc/standalone/wasi);
`tsc --noEmit` -> 0; `tests/issue-3274-slice3.test.ts` (3) green — __object_create
+ __getPrototypeOf reachable + correct under --target standalone, zero host imports.

Subtask of #3182. Stacked on #3076 (slice 1) via slice 2 — do not enqueue until
slices 1 and 2 land.

✓
